make severity its own application
