+++ chirps/base_app/management/commands/initialize_app.py
@@ -68,6 +68,8 @@ class Command(BaseCommand):
 
         # Run the 'loaddata' command
         self.stdout.write(self.style.WARNING('Loading data from fixtures...'))
+        self.load_data_from_fixtures('/severity/fixtures/severity')
+        self.stdout.write(self.style.SUCCESS('Severity data loaded from fixtures'))
         self.load_data_from_fixtures('/policy/fixtures/policy')
         self.stdout.write(self.style.SUCCESS('Policy data loaded from fixtures'))
         self.load_data_from_fixtures('/embedding/fixtures/embedding')
