+++ chirps/policy/templates/policy/dashboard.html
@@ -54,7 +54,7 @@
 
     <div class="tab-pane fade" id="tabs-3" role="tabpanel" aria-labelledby="tab-3">
         <div class="accordion" id="template-panel-accordion">
-            {% include 'policy/severity_management.html' with severities=severities %}
+            {% include 'severity/severity_management.html' with severities=severities %}
         </div>
     </div>
