+++ chirps/policy/views.py
@@ -1,17 +1,13 @@
 """Views for the policy app."""
 from django.contrib import messages
-from django.contrib.auth.decorators import login_required, user_passes_test
-from django.http import HttpResponse, HttpResponseNotAllowed
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 from django.shortcuts import get_object_or_404, redirect, render
 from django.views.decorators.http import require_http_methods
+from severity.forms import CreateSeverityForm, EditSeverityForm
 
-from .forms import CreateSeverityForm, EditSeverityForm, PolicyForm
+from .forms import PolicyForm
 from .models import Policy, PolicyVersion, Rule, Severity
-
-
-def is_superuser(user):
-    """Check if the user is a superuser."""
-    return user.is_superuser
 
 
 @login_required
@@ -195,53 +191,5 @@ def archive(request, policy_id):
 
     # Add an info message
     messages.info(request, 'Policy has been archived.')
-
-    return redirect('policy_dashboard')
-
-
-@user_passes_test(is_superuser)
-@login_required
-def create_severity(request):
-    """Create a new severity."""
-    if request.method == 'POST':
-        form = CreateSeverityForm(request.POST)
-        if form.is_valid():
-            form.save()
-            messages.success(request, 'Severity created successfully.')
-        else:
-            messages.error(request, 'Error creating severity.')
-        return redirect('policy_dashboard')
-
-    return HttpResponseNotAllowed(['POST'])
-
-
-@user_passes_test(is_superuser)
-@login_required
-def edit_severity(request, severity_id):
-    """Edit an existing severity."""
-    severity = get_object_or_404(Severity, id=severity_id)
-    if request.method == 'POST':
-        form = EditSeverityForm(request.POST, instance=severity)
-        if form.is_valid():
-            form.save()
-            messages.success(request, 'Severity updated successfully.')
-        else:
-            messages.error(request, 'Error updating severity.')
-        return redirect('policy_dashboard')
-
-    return HttpResponseNotAllowed(['POST'])
-
-
-@user_passes_test(is_superuser)
-@login_required
-def archive_severity(request, severity_id):
-    """Archive a severity."""
-    try:
-        severity = Severity.objects.get(id=severity_id)
-        severity.archived = True
-        severity.save()
-        messages.success(request, f"Severity '{severity.name}' has been archived.")
-    except Severity.DoesNotExist:
-        messages.error(request, 'Severity not found.')
 
     return redirect('policy_dashboard')
