+++ chirps/severity/apps.py
@@ -1,0 +1,8 @@
+from django.apps import AppConfig
+
+
+class SeverityConfig(AppConfig):
+    """Severity app config"""
+
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'severity'
