+++ chirps/severity/forms.py
@@ -1,0 +1,15 @@
+from django import forms
+from severity.models import Severity
+
+
+class CreateSeverityForm(forms.ModelForm):
+    """Form for creating a new severity."""
+
+    class Meta:
+        model = Severity
+        fields = ['name', 'value', 'color']
+        widgets = {'color': forms.widgets.TextInput(attrs={'type': 'color'})}
+
+
+class EditSeverityForm(CreateSeverityForm):
+    """Form for editing an existing severity."""
