+++ chirps/severity/templates/severity/severity_management.html
@@ -1,4 +1,4 @@
-{% load policy_filters %}
+{% load severity_filters %}
 
 <div class="container">
     <table class="table">
