+++ chirps/severity/templatetags/severity_filters.py
@@ -5,6 +5,6 @@ register = template.Library()
 
 
 @register.filter
-def get_item(dictionary, key):
+def get_item(dictionary: dict, key: str) -> str | None:
     """Return the given key from a dictionary."""
     return dictionary.get(key)
