+++ mypy.ini
@@ -1,15 +1,13 @@
 [mypy]
+mypy_path = chirps
 explicit_package_bases = True
 namespace_packages = True
 follow_imports = silent
 disallow_untyped_defs = True
-exclude = .*/tests\.py$|.*migrations.*|chirps/account.*|chirps/asset.*|chirps/base_app.*|chirps/chirps.*|chirps/embedding.*|chirps/policy.*|chirps/scan.*
+exclude = .*/tests\.py$|.*/models\.py$|.*migrations.*|chirps/account.*|chirps/asset.*|chirps/base_app.*|chirps/chirps.*|chirps/embedding.*|chirps/policy.*|chirps/scan.*
 
 [mypy-mantium_client.*]
 ignore_missing_imports = True
 
 [mypy-mantium_spec.*]
-ignore_missing_imports = True
-
-[mypy-chirps.*]
 ignore_missing_imports = True
